+++ core/src/main/scala/com/krux/hyperion/activity/MapReduceStep.scala
@@ -1,6 +1,7 @@
 package com.krux.hyperion.activity
 
 import com.krux.hyperion.adt.{ HString, HS3Uri }
+import com.krux.hyperion.common.Escapable
 
 /**
  * A MapReduce step that runs on MapReduce Cluster
@@ -9,12 +10,14 @@ case class MapReduceStep private (
   jarUri: HString,
   mainClass: Option[MainClass],
   args: Seq[HString]
-) {
+) extends Escapable {
 
   def withMainClass(mainClass: MainClass) = copy(mainClass = Option(mainClass))
   def withArguments(arg: HString*) = copy(args = args ++ arg)
 
-  def serialize: String = (jarUri +: mainClass.map(_.toString).toSeq ++: args).mkString(",")
+  def serialize: String = (jarUri +: mainClass.map(_.toString).toSeq ++: args)
+    .map(x => escape(x.toString, ','))
+    .mkString(",")
 
   override def toString = serialize